Fix unit test dealing with walletd
