+++ polling/coincap_test.go
@@ -6,11 +6,14 @@ import (
 
 	"github.com/pegnet/pegnet/common"
 	. "github.com/pegnet/pegnet/polling"
+	"github.com/pegnet/pegnet/testutils"
 	"github.com/zpatrick/go-config"
 )
 
 // TestCoinCapPeggedAssets tests all the crypto assets are found on coinmarket cap
 func TestCoinCapPeggedAssets(t *testing.T) {
+	defer func() { http.DefaultClient = &http.Client{} }() // Don't leave http broken
+
 	c := config.NewConfig([]config.Provider{common.NewUnitTestConfigProvider()})
 	peg := make(PegAssets)
 
@@ -30,7 +33,7 @@ func TestFixedCoinCapRatesPeggedAssets(t *testing.T) {
 	c := config.NewConfig([]config.Provider{common.NewUnitTestConfigProvider()})
 
 	// Set default http client to return what we expect from apilayer
-	cl := GetClientWithFixedResp([]byte(coincapResp))
+	cl := testutils.GetClientWithFixedResp([]byte(coincapResp))
 	http.DefaultClient = cl
 
 	peg := make(PegAssets)
